refactor(cockpit): extract inbox + worker roster panels (#405)
Move the inbox renderers (``render_inbox_panel``, ``_render_inbox_panel``,
``_inbox_db_sources``, ``_count_inbox_tasks_for_label``,
``_render_work_service_issues``) and the worker-roster gather/render
helpers (``WorkerRosterRow``, ``_worker_roster_sort_key``,
``_format_worker_turn_label``, ``_last_commit_age``,
``_gather_worker_roster``, ``_try_load_supervisor_for_config``,
``_render_worker_roster_panel``, ``_register_worker_roster_rail_item``,
``_gather_activity_feed``) out of ``cockpit.py`` into a new
``pollypm.cockpit_inbox`` module that follows the same
contract-comment + lazy-Textual-import shape as ``cockpit_activity.py``.

``cockpit.py`` keeps the cross-section orchestrators
(``build_cockpit_detail`` + ``_build_cockpit_detail_inner`` +
``_build_cockpit_detail_dispatch``) plus the metrics + palette layers
that still depend on multiple cockpit subsystems. The new module is
re-exported from ``cockpit.py`` so legacy importers
(``from pollypm.cockpit import _count_inbox_tasks_for_label``,
``WorkerRosterRow``, …) and the tests/plugins that reach for them keep
working without churn.

Net effect: ``cockpit.py`` drops from 3873 LOC (pre-#403) to 1119 LOC.
The 175-test cockpit/rail/inbox/worker-roster/metrics/palette suite
passes locally.
